make parameter name check identical to encoded value check
