+++ core/src/main/java/com/graphhopper/routing/ev/IntEncodedValueImpl.java
@@ -19,8 +19,7 @@ package com.graphhopper.routing.ev;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
-import javax.lang.model.SourceVersion;
+import com.graphhopper.util.Helper;
 
 /**
  * Implementation of the IntEncodedValue via a certain number of bits (that determines the maximum value) and
@@ -70,7 +69,7 @@ public class IntEncodedValueImpl implements IntEncodedValue {
      * @param storeTwoDirections     true if forward and backward direction of the edge should get two independent values.
      */
     public IntEncodedValueImpl(String name, int bits, int minStorableValue, boolean negateReverseDirection, boolean storeTwoDirections) {
-        if (!isValidEncodedValue(name))
+        if (!Helper.isValidEncodedValue(name))
             throw new IllegalArgumentException("EncodedValue name wasn't valid: " + name + ". Use lower case letters, underscore and numbers only.");
         if (bits <= 0)
             throw new IllegalArgumentException(name + ": bits cannot be zero or negative");
@@ -233,29 +232,5 @@ public class IntEncodedValueImpl implements IntEncodedValue {
     @Override
     public final String toString() {
         return getName();
-    }
-
-    static boolean isValidEncodedValue(String name) {
-        if (name.length() < 2 || name.startsWith("in_") || name.startsWith("backward_")
-                || !isLowerLetter(name.charAt(0)) || SourceVersion.isKeyword(name))
-            return false;
-
-        int underscoreCount = 0;
-        for (int i = 1; i < name.length(); i++) {
-            char c = name.charAt(i);
-            if (c == '_') {
-                if (underscoreCount > 0) return false;
-                underscoreCount++;
-            } else if (!isLowerLetter(c) && !Character.isDigit(c)) {
-                return false;
-            } else {
-                underscoreCount = 0;
-            }
-        }
-        return true;
-    }
-
-    private static boolean isLowerLetter(char c) {
-        return c >= 'a' && c <= 'z';
     }
 }
